transfere responsabilidade da mainActivity para o fragment
