+++ Calculadora/.idea/misc.xml
@@ -4,7 +4,9 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout-port/activity_main.xml" value="0.20416666666666666" />
-        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout/activity_main.xml" value="0.5" />
+        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout-port/fragment_calculator.xml" value="0.36666666666666664" />
+        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout/activity_main.xml" value="0.1703125" />
+        <entry key="..\:/Users/gustavo/Documents/Studio/GitHub/Calculadora/Calculadora/app/src/main/res/layout/fragment_calculator.xml" value="0.25" />
       </map>
     </option>
   </component>
